Mypy et al.

+++ micromechanics/indentation/hertz.py
@@ -46,12 +46,11 @@ class IndentationHertzMixin:
     fitMask[np.argmax(self.p):] = False
     if np.count_nonzero(fitMask)<3:
       raise ValueError("hertzFit: not enough data points in forceRange before maximum load")
-    xFit = np.asarray(self.h[fitMask], dtype=np.float64)
-    yFit = np.asarray(self.p[fitMask], dtype=np.float64)
+    xFit = np.asarray(self.h[fitMask], dtype=float)
+    yFit = np.asarray(self.p[fitMask], dtype=float)
     depthRange = (float(xFit.min()), float(xFit.max()))
-    para0 = np.array([0.0, 5000.0], dtype=np.float64)
-    bounds = (np.array([-depthRange[0], 0.0], dtype=np.float64),
-              np.array([depthRange[0], 50000.0], dtype=np.float64))
+    para0 = np.array([0.0, 5000.0], dtype=float)
+    bounds = (np.array([-depthRange[0], 0.0], dtype=float), np.array([depthRange[0], 50000.0], dtype=float))
     fitElast, _ = curve_fit(hertzEquation, xFit, yFit, p0=para0, bounds=bounds) # pylint: disable=unbalanced-tuple-unpacking
     if self.output['verbose']>1:
       print('Depth range', depthRange)
@@ -118,7 +117,7 @@ class IndentationHertzMixin:
 
     def funct(depth:np.ndarray, prefactor:float, h0:float) -> np.ndarray:
       diff           = depth-h0
-      if isinstance(diff, np.float64):
+      if isinstance(diff, float):
         diff = max(diff,0.0)
       else:
         diff[diff<0.0] = 0.0
